csrf pour get

+++ src/Controller/AdminController.php
@@ -8,6 +8,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * @IsGranted("ROLE_ADMIN")
@@ -31,13 +32,22 @@ class AdminController extends AbstractController
      * @Route("/admin/delete/{userId}", name="app_admin_delete")
      * Permet de supprimer un utilisateur
      */
-    public function deleteUser(UserRepository $ur, int $userId, Helpers $helper): Response
+    public function deleteUser(UserRepository $ur, int $userId, Helpers $helper, Request $request): Response
     {
         $user = $ur->find($userId);
-        $helper->deleteDir($this->getParameter('app.imageDirectory').'/'.$user->getUsername());
-        $ur->remove($user);
-        $this->addFlash('success', 'L\'utilisateur a été supprimé');
 
+        $submittedToken = $request->get('csrf_token');
+
+        if ($this->isCsrfTokenValid('remove', $submittedToken))
+        {
+            $helper->deleteDir($this->getParameter('app.imageDirectory').'/'.$user->getUsername());
+            $ur->remove($user);
+            $this->addFlash('success', 'L\'utilisateur a été supprimé');
+        }
+        else{
+            return $helper->error(403, 'Erreur 403','Vous n\'avez pas pu être authentifié comme l\'auteur de cette action');
+        }
+    
         return $this->redirectToRoute('app_admin');
     }
 }
